Move DataConversionException to top level exception package
