+++ src/test/java/seedu/address/commons/util/ConfigUtilTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import seedu.address.commons.core.Config;
-import seedu.address.commons.exceptions.DataConversionException;
+import seedu.address.exception.DataConversionException;
 
 public class ConfigUtilTest {
