+++ src/test/java/seedu/address/storage/JsonUserPrefsStorageTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.commons.exceptions.DataConversionException;
+import seedu.address.exception.DataConversionException;
 import seedu.address.model.UserPrefs;
 
 public class JsonUserPrefsStorageTest {
